Add python interpreter check and transfer
This change checks the value of the provided python interpreter to see
if it contains an operating system separator. If it does the default
behaviour to parse the given string is executed and retunred. If no
separator is found the string will now be returned as-is and assumed to
be a templated option which ansible internals will handle natively.

This also adds the `mitogen_python_interpreter` which will default to
the `ansible_python_interpreter` when unset. This will allow deployers
to set python interpreters for mitogen as a constant which is important
as the python interpreter may also be set at the task level.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>